Moving metrics service to lib/
