+++ web-local/src/lib/components/metrics-definition/MetricsDefinitionExploreMetricsButton.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
   import { goto } from "$app/navigation";
-  import { BehaviourEventMedium } from "@rilldata/web-local/common/metrics-service/BehaviourEventTypes";
+  import { BehaviourEventMedium } from "@rilldata/web-local/lib/metrics/service/BehaviourEventTypes";
   import {
     MetricsEventScreenName,
     MetricsEventSpace,
-  } from "@rilldata/web-local/common/metrics-service/MetricsTypes";
+  } from "@rilldata/web-local/lib/metrics/service/MetricsTypes";
   import { Button } from "../button";
   import ExploreIcon from "../icons/Explore.svelte";
   import Tooltip from "../tooltip/Tooltip.svelte";
